requirements.txt
bleak==0.22.2
contourpy==1.2.1
cycler==0.12.1
dbus-fast==2.21.3
fonttools==4.53.0
kiwisolver==1.4.5
matplotlib==3.9.0
MouseInfo==0.1.3
numpy==2.0.0
packaging==24.1
pillow==10.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.1.2
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
python-uinput==1.0.1
python3-xlib==0.15
pytweening==1.2.0
six==1.16.0
